Add a workflow to update PM files after lock file
Renovate automatically creates PRs to update lock files. The most common PR
failure after this is that Project Manager-generated files are out of date due
to upstream changes. Renovate doesn’t allow us to run any commands on its PR, so
this runs them in a new workflow, and then pushes any changes to the PR.